Add JSON config, migration tools, metadata

+++ augur/git.py
@@ -9,6 +9,7 @@ import shutil
 import pandas as pd
 from dulwich.repo import Repo
 from dulwich import porcelain as git
+from augur import logger
 
 # end imports
 # (don't remove the above line, it's for a script)
@@ -27,6 +28,10 @@ class Git(object):
         self.__repo_urls = list_of_repositories
         self.__repos = {}
         self.__folder = os.path.abspath(storage_folder)
+        if not os.path.exists(self.__folder):
+            os.makedirs(self.__folder)
+        self.__log_location = os.path.join(self.__folder, 'git_log.txt')
+        self.__log = open(self.__log_location, 'wb')
         self.sync()
 
     def get_repo_object(self, repo_url):
@@ -36,31 +41,47 @@ class Git(object):
         :param listOfRepositories: List of URLs to Git repos to analyze
         :param storageFolder: Folder to download the repositories to
         """
-        if hasattr(self.__repos, repo_url):
+        if repo_url in self.__repos:
             return self.__repos[repo_url]
         else:
-            raise ValueError("{} is not in the list_of_repositories".format(repo_url))
+            try:
+                folder_name = os.path.splitext(os.path.basename(repo_url))[0]
+                repo_path = os.path.join(self.__folder, folder_name)
+                self.__repos[repo_url] = Repo(repo_path)
+                return self.__repos[repo_url]
+            except:
+                raise ValueError("{} is not in the list_of_repositories".format(repo_url))
 
     def sync(self):
         """
         Makes sure the storage_folder contains updated versions of all the repos
         """
-        for repo_url in self.__repo_urls:
-            folder_name = os.path.splitext(os.path.basename(repo_url))[0]
-            repo_path = os.path.join(self.__folder, folder_name)
-            if not os.path.exists(repo_path):
-                git.clone(repo_url, repo_path)
-            repo = Repo(repo_path)
-            try:
-                git.pull(repo)
-            except:
-                shutil.rmtree(repo_path)
-                git.clone(repo_url, repo_path)
+        lockfile_path = os.path.join(self.__folder, 'lock')
+        if not os.path.exists(lockfile_path):
+            lockfile = open(lockfile_path, 'w')
+            for repo_url in self.__repo_urls:
+                logger.info('Git: Updating %s', repo_url)
+                folder_name = os.path.splitext(os.path.basename(repo_url))[0]
+                repo_path = os.path.join(self.__folder, folder_name)
+                if not os.path.exists(repo_path):
+                    logger.debug('Cloning %s', repo_url)
+                    git.clone(repo_url, repo_path, outstream=self.__log, errstream=self.__log)
                 repo = Repo(repo_path)
-            self.__repos[repo_url] = repo
+                try:
+                    logger.debug('Pulling %s', repo_url)
+                    git.pull(repo, remote_location=repo_url, outstream=self.__log, errstream=self.__log)
+                except Exception as e:
+                    logger.debug('Re-Cloning %s because %s', repo_url, str(e))
+                    shutil.rmtree(repo_path)
+                    git.clone(repo_url, repo_path, outstream=self.__log, errstream=self.__log)
+                    repo = Repo(repo_path)
+                self.__repos[repo_url] = repo
+            lockfile.close()
+            os.remove(lockfile_path)
 
-    def lines_changed_minus_whitespace(self):
+    def lines_changed_minus_whitespace(self, repo_url):
         """
         Makes sure the storageFolder contains updated versions of all the repos
         """
-        return []
+        repo = self.get_repo_object(repo_url)
+        return repo
